implemented rudementary shroomie
Shroomie lacks the ability to shoot towards enemies, but is currently capable of facing the target enemy to be attacked. It should be noted that the current lockon system does not guarantee prioritization of the enemy furthest along the path. A flipping animation has been implemented for when Shroomie changes her direction of attack.