projectiles
Implemented projectiles. Also I'm pretty sure the targetting inherently prioritizes the bullet closest to the end of the path but I could be wrong and will have to continue to keep an eye out for the behaviour. Worst comes to worst I just have to expand on the targetting script which shouldn't be a big deal.